Handle redirecting login
