+++ packages/admin-panel/src/api/mutations/useLogout.js
@@ -4,17 +4,14 @@
  */
 
 import { useMutation, useQueryClient } from 'react-query';
-import { useNavigate } from 'react-router';
 import { post } from '../../VizBuilderApp/api';
 
 export const useLogout = () => {
   const queryClient = useQueryClient();
-  const navigate = useNavigate();
 
   return useMutation('logout', () => post('logout'), {
     onSuccess: () => {
       queryClient.invalidateQueries();
-      navigate('/login');
     },
   });
 };
