+++ packages/admin-panel/src/pages/LoginPage.jsx
@@ -4,11 +4,11 @@
  */
 
 import React from 'react';
-import PropTypes from 'prop-types';
-import { Navigate, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
-import { LoginForm } from '../authentication/LoginForm';
-import { useUser } from '../VizBuilderApp/api';
+import { useForm } from 'react-hook-form';
+import { Link } from '@material-ui/core';
+import { LoginForm } from '@tupaia/ui-components';
+import { useLogin } from '../api/mutations';
 
 const Container = styled.section`
   display: flex;
@@ -21,26 +21,23 @@ const Container = styled.section`
   }
 `;
 
-export const LoginPage = ({ redirectTo }) => {
-  const location = useLocation();
-  const { isLoggedIn, isLoading } = useUser();
-  if (isLoading) return null;
-  if (isLoggedIn) {
-    const redirectUrl = redirectTo || location.state?.from || '/';
-    return <Navigate to={redirectUrl} />;
-  }
-
+export const LoginPage = () => {
+  const formContext = useForm();
+  const { mutate: onLogin, isLoading, error } = useLogin();
+  const requestAnAccountUrl = 'https://info.tupaia.org/contact';
   return (
     <Container>
-      <LoginForm />
+      <LoginForm
+        onSubmit={onLogin}
+        isLoading={isLoading}
+        error={error}
+        formContext={formContext}
+        RegisterLinkComponent={
+          <Link href={requestAnAccountUrl} target="_blank">
+            Request an account here
+          </Link>
+        }
+      />
     </Container>
   );
-};
-
-LoginPage.propTypes = {
-  redirectTo: PropTypes.string,
-};
-
-LoginPage.defaultProps = {
-  redirectTo: null,
 };
